Output a Game Gear compatible palette
Note that this does not support the extra colours of the Game Gear. It
simply outputs a 12-bit copy of the palette so that tiles made for the
Master System can also be used on the Game Gear.

The application including the generated file will need to define one
of TARGET_SMS or TARGET_GG to get the appropriate palette array.

While updating the generated files, we also make the arrays static, as
the symbols don't need to be reachable from other objects.