Tile de-duplication
For now this is restricted to 8x8 tiles within a single file.

A hard-coded limit is currently in place of 512 unique tiles before
no more can be added to the list. Using all of this would fill all of
the vram in the Master System, leaving no space for other tables.