Support for sprite panels
Often, we want to use game objects that are larger than a single 8x8
tile. Especially with de-duplication in place, it is nice for this to
be somewhat automated.

This patch introduces a --panels option, which allows a per-image
panel size and count to be described. When present, this option results
in an array being added to pattern_index for each paneled file.
The array is indexed by the panel number, and each row contains the
full list of indexes for that panel.

For now, this feature is restricted to mode-4 with de-duplication.

To tame the number of parameters some functions need, global state is
added to keep track of the dimensions of the current image file.