Option to reserve patterns (mode-4 only for now)
For background patterns in mode-4, there are 512 indexes that can be
represented, allowing background patterns to take up any space in VRAM
not used by other tables. For sprites however, only the low 256 indexes
may be used.

One way of working around this is for patterns to be copied into a low
index at run-time. To allow for this, we need the ability to reserve
some pattern space early on.

The "--reserve <name,n>" option is added to reserve tiles. It can occur
multiple times on the command line, and can occur between image files.