Separate pattern array for each input file
This removes the need to store a bunch of zeros between patterns if
some area needs to be reserved for dynamic patterns.

The pattern indexing provided in the panel arrays indexes the pattern
arrays directly. This means that it is easy for the CPU to interact
with them, but other than the first they cannot directly be put into
the VDP. As the indexes cannot go directly into the VDP any more, the
palette-bit is also removed from the index.

This is something that may be revisited in the future, but a flexible
method of defining the VRAM-address of the pattern arrays will need to
be worked out.

The definitions providing the first index for each input file have also
been removed, as the different arrays can now be accessed by name.